temp-to-be-deleted/multi-agents/code-reviewers/pyproject.toml
[project]
name = "code_review"
version = "0.1.0"
description = "Code Review Agents"
authors = [{name = "Your Name", email = "you@example.com"}]
readme = "README.md"
requires-python = ">=3.12,<=3.13"
dependencies = [
    "crewai[tools]>=0.152.0",
    "python-dotenv>=1.0.1",
    "html2text>=2024.2.26",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
]

[project.scripts]
code_review = "code_review.main:run"
train = "code_review.main:train"